# ENH: Implement sindex.query_bulk in overlay