@adriangb 2020-05-16
I ended up implementing a more optimized version based on my comments above. The only thing I don't like is the `if compat.USE_PYGEOS` logic.

Benchmarks:
| before      | after       | ratio | test                                                   | pygeos installed? |
|-------------|-------------|-------|--------------------------------------------------------|-------------------|
| 328±80ms    | 106±0.3ms   | 0.32  | overlay.Countries.time_overlay('difference')           | Y                 |
| 327±80ms    | 244±70ms    | ~0.75 | overlay.Countries.time_overlay('difference')           | N                 |
| 724±20ms    | 415±70ms    | 0.57  | overlay.Countries.time_overlay('identity')             | Y                 |
| 805±20ms    | 624±20ms    | 0.78  | overlay.Countries.time_overlay('identity')             | N                 |
| 181±20ms    | 124±0.3ms   | 0.68  | overlay.Countries.time_overlay('intersection')         | Y                 |
| 261±5ms     | 227±3ms     | 0.87  | overlay.Countries.time_overlay('intersection')         | N                 |
| 587±40ms    | 269±5ms     | 0.46  | overlay.Countries.time_overlay('symmetric_difference') | Y                 |
| 573±10ms    | 599±200ms   | 1.05  | overlay.Countries.time_overlay('symmetric_difference') | N                 |
| 752±100ms   | 376±3ms     | 0.50  | overlay.Countries.time_overlay('union')                | Y                 |
| 791±7ms     | 628±6ms     | 0.79  | overlay.Countries.time_overlay('union')                | N                 |
| 10.00±0.4ms | 5.21±0.09ms | 0.52  | overlay.Small.time_overlay('difference')               | Y                 |
| 9.72±0.3ms  | 7.29±0.3ms  | 0.75  | overlay.Small.time_overlay('difference')               | N                 |
| 42.1±5ms    | 25.2±0.3ms  | 0.60  | overlay.Small.time_overlay('identity')                 | Y                 |
| 42.5±5ms    | 30.3±2ms    | 0.71  | overlay.Small.time_overlay('identity')                 | N                 |
| 18.7±3ms    | 12.2±0.2ms  | 0.65  | overlay.Small.time_overlay('intersection')             | Y                 |
| 18.7±3ms    | 15.0±3ms    | ~0.80 | overlay.Small.time_overlay('intersection')             | N                 |
| 28.4±3ms    | 16.5±0.4ms  | 0.58  | overlay.Small.time_overlay('symmetric_difference')     | Y                 |
| 28.2±3ms    | 20.4±0.7ms  | 0.72  | overlay.Small.time_overlay('symmetric_difference')     | N                 |
| 42.3±5ms    | 24.7±0.3ms  | 0.58  | overlay.Small.time_overlay('union')                    | Y                 |
| 41.1±6ms    | 31.2±4ms    | ~0.76 | overlay.Small.time_overlay('union')                    | N                 |