@adriangb 2020-05-16
I think this is ready for review, pinging @jorisvandenbossche @martinfleis and @brendan-ward.

I can always revert the vectorization changes I made to minimize the diff, let me know.