@adriangb 2020-05-16
Benchmarks for 5c1187c:

| before      | after      | ratio | test                                                   | pygeos installed? |
|-------------|------------|-------|--------------------------------------------------------|-------------------|
| 238±2ms     | 175±4ms    | 0.73  | overlay.Countries.time_overlay('difference')           | Y                 |
| 259±7ms     | 219±2ms    | 0.85  | overlay.Countries.time_overlay('difference')           | N                 |
| 658±10ms    | 579±60ms   | ~0.88 | overlay.Countries.time_overlay('identity')             | Y                 |
| 748±2ms     | 693±50ms   | 0.93  | overlay.Countries.time_overlay('identity')             | N                 |
| 162±3ms     | 163±20ms   | 1.01  | overlay.Countries.time_overlay('intersection')         | Y                 |
| 252±6ms     | 227±1ms    | ~0.90 | overlay.Countries.time_overlay('intersection')         | N                 |
| 507±2ms     | 418±10ms   | ~0.82 | overlay.Countries.time_overlay('symmetric_difference') | Y                 |
| 539±9ms     | 507±10ms   | 0.94  | overlay.Countries.time_overlay('symmetric_difference') | N                 |
| 663±5ms     | 543±20ms   | ~0.82 | overlay.Countries.time_overlay('union')                | Y                 |
| 740±5ms     | 693±7ms    | 0.94  | overlay.Countries.time_overlay('union')                | N                 |
| 8.39±0.2ms  | 8.95±5ms   | 1.07  | overlay.Small.time_overlay('difference')               | Y                 |
| 8.57±0.05ms | 8.84±0.4ms | 1.03  | overlay.Small.time_overlay('difference')               | N                 |
| 35.0±0.5ms  | 33.9±0.7ms | 0.97  | overlay.Small.time_overlay('identity')                 | Y                 |
| 35.8±0.9ms  | 33.6±0.2ms | 0.94  | overlay.Small.time_overlay('identity')                 | N                 |
| 14.9±0.6ms  | 13.9±0.7ms | 0.94  | overlay.Small.time_overlay('intersection')             | Y                 |
| 14.9±0.08ms | 13.2±1ms   | ~0.89 | overlay.Small.time_overlay('intersection')             | N                 |
| 23.5±0.6ms  | 38.1±20ms  | ~1.62 | overlay.Small.time_overlay('symmetric_difference')     | Y                 |
| 23.7±0.5ms  | 23.7±0.3ms | 1.00  | overlay.Small.time_overlay('symmetric_difference')     | N                 |
| 34.6±0.5ms  | 33.6±0.9ms | 0.97  | overlay.Small.time_overlay('union')                    | Y                 |
| 35.5±0.3ms  | 32.7±0.4ms | 0.92  | overlay.Small.time_overlay('union')                    | N                 |