@adriangb 2020-05-16
Keeping track of TODOs stemming from this PR that I will open follow up issues for:

- [ ] A better method to test/compare MultiPolygons

- [ ] Vectorization/performance improvements for _overlay_difference