@adriangb 2020-05-16
@martinfleis could you please restart this job? It looks like a CI failure: https://travis-ci.org/github/geopandas/geopandas/jobs/687802733