@adriangb 2020-05-16
Final benchmarks:

<details>
<summary>e38584eda0db38f3af9208b0726b2bed4da6bd82 (current)</summary>
| before     | after      | ratio | test                                                   | pygeos installed? |
|------------|------------|-------|--------------------------------------------------------|-------------------|
| 306±10ms   | 160±4ms    | 0.52  | overlay.Countries.time_overlay('difference')           | Y                 |
| 300±4ms    | 286±10ms   | 0.95  | overlay.Countries.time_overlay('difference')           | N                 |
| 859±30ms   | 510±30ms   | 0.59  | overlay.Countries.time_overlay('identity')             | Y                 |
| 937±100ms  | 961±100ms  | 1.03  | overlay.Countries.time_overlay('identity')             | N                 |
| 184±2ms    | 128±20ms   | 0.69  | overlay.Countries.time_overlay('intersection')         | Y                 |
| 266±7ms    | 278±60ms   | 1.04  | overlay.Countries.time_overlay('intersection')         | N                 |
| 613±40ms   | 401±30ms   | 0.65  | overlay.Countries.time_overlay('symmetric_difference') | Y                 |
| 632±10ms   | 704±50ms   | ~1.11 | overlay.Countries.time_overlay('symmetric_difference') | N                 |
| 796±10ms   | 575±70ms   | 0.72  | overlay.Countries.time_overlay('union')                | Y                 |
| 881±90ms   | 793±100ms  | ~0.90 | overlay.Countries.time_overlay('union')                | N                 |
| 9.63±0.6ms | 8.48±1ms   | ~0.88 | overlay.Small.time_overlay('difference')               | Y                 |
| 12.6±1ms   | 14.4±0.9ms | ~1.14 | overlay.Small.time_overlay('difference')               | N                 |
| 42.7±3ms   | 35.6±3ms   | ~0.84 | overlay.Small.time_overlay('identity')                 | Y                 |
| 42.0±3ms   | 42.3±1ms   | 1.01  | overlay.Small.time_overlay('identity')                 | N                 |
| 17.3±3ms   | 12.8±2ms   | ~0.74 | overlay.Small.time_overlay('intersection')             | Y                 |
| 22.7±2ms   | 18.4±4ms   | ~0.81 | overlay.Small.time_overlay('intersection')             | N                 |
| 27.3±1ms   | 22.0±0.7ms | ~0.81 | overlay.Small.time_overlay('symmetric_difference')     | Y                 |
| 29.6±1ms   | 30.9±5ms   | 1.04  | overlay.Small.time_overlay('symmetric_difference')     | N                 |
| 41.1±2ms   | 29.3±1ms   | 0.71  | overlay.Small.time_overlay('union')                    | Y                 |
| 45.7±3ms   | 42.6±5ms   | 0.93  | overlay.Small.time_overlay('union')                    | N                 |
</details>

<details>
<summary>e9bb8f2540597b862b94a7d7d903ad03ef23696a (vectorized)</summary>

| before     | after       | ratio | test                                                   | pygeos installed? |
|------------|-------------|-------|--------------------------------------------------------|-------------------|
| 306±10ms   | 111±7ms     | 0.36  | overlay.Countries.time_overlay('difference')           | Y                 |
| 300±4ms    | 188±20ms    | 0.63  | overlay.Countries.time_overlay('difference')           | N                 |
| 859±30ms   | 430±30ms    | 0.50  | overlay.Countries.time_overlay('identity')             | Y                 |
| 937±100ms  | 702±50ms    | ~0.75 | overlay.Countries.time_overlay('identity')             | N                 |
| 184±2ms    | 127±2ms     | 0.69  | overlay.Countries.time_overlay('intersection')         | Y                 |
| 266±7ms    | 224±10ms    | ~0.84 | overlay.Countries.time_overlay('intersection')         | N                 |
| 613±40ms   | 292±30ms    | 0.48  | overlay.Countries.time_overlay('symmetric_difference') | Y                 |
| 632±10ms   | 411±20ms    | 0.65  | overlay.Countries.time_overlay('symmetric_difference') | N                 |
| 796±10ms   | 398±10ms    | 0.50  | overlay.Countries.time_overlay('union')                | Y                 |
| 881±90ms   | 765±90ms    | ~0.87 | overlay.Countries.time_overlay('union')                | N                 |
| 9.63±0.6ms | 6.25±0.7ms  | ~0.65 | overlay.Small.time_overlay('difference')               | Y                 |
| 12.6±1ms   | 7.11±1ms    | 0.56  | overlay.Small.time_overlay('difference')               | N                 |
| 42.7±3ms   | 28.4±9ms    | ~0.67 | overlay.Small.time_overlay('identity')                 | Y                 |
| 42.0±3ms   | 29.6±2ms    | 0.71  | overlay.Small.time_overlay('identity')                 | N                 |
| 17.3±3ms   | 11.9±0.03ms | 0.69  | overlay.Small.time_overlay('intersection')             | Y                 |
| 22.7±2ms   | 13.3±0.8ms  | 0.58  | overlay.Small.time_overlay('intersection')             | N                 |
| 27.3±1ms   | 17.2±2ms    | 0.63  | overlay.Small.time_overlay('symmetric_difference')     | Y                 |
| 29.6±1ms   | 19.7±3ms    | ~0.67 | overlay.Small.time_overlay('symmetric_difference')     | N                 |
| 41.1±2ms   | 30.3±8ms    | ~0.74 | overlay.Small.time_overlay('union')                    | Y                 |
| 45.7±3ms   | 28.6±2ms    | 0.62  | overlay.Small.time_overlay('union')                    | N                 |
</details>